fix builds

+++ test/hotspot/gtest/opto/test_rangeinference.cpp
@@ -54,9 +54,9 @@ static void test_canonicalize_constraints_trivial() {
   ASSERT_FALSE(TypeInt::CC_NE->contains(0));
   ASSERT_FALSE(TypeInt::CC_NE->contains(-2));
   ASSERT_FALSE(TypeInt::CC_NE->contains(2));
-  ASSERT_FALSE(TypeLong::NON_ZERO->contains(0L));
-  ASSERT_TRUE(TypeLong::NON_ZERO->contains(1L));
-  ASSERT_TRUE(TypeLong::NON_ZERO->contains(-1L));
+  ASSERT_FALSE(TypeLong::NON_ZERO->contains(jlong(0)));
+  ASSERT_TRUE(TypeLong::NON_ZERO->contains(jlong(1)));
+  ASSERT_TRUE(TypeLong::NON_ZERO->contains(jlong(-1)));
 }
 
 template <class S, class U>
